bugfix: build with -tags go_tarantool_ssl_disable
The patch fixes build with the build tag `go_tarantool_ssl_disable`:

1. It moves tests with OpenSslDialer to a test file that executes only
with the tag.
2. It defines structure `sslOpts` in the common place to use it in
the code with/without the flag.

Finally, it adds tests to CI with the build tag.

Closes #357
